refactor : suppression des différents packages superflux pour l'instant et modification nom application
