--- src/main/java/com/isn/quizplatform/dao/QuizDao.java
@@ -1,5 +1,0 @@
-package com.isn.quizplatform.dao;
-
-public class QuizDao {
-
-}
