--- src/main/java/com/isn/quizplatform/dto/QuizDto.java
@@ -1,5 +1,0 @@
-package com.isn.quizplatform.dto;
-
-public class QuizDto {
-
-}
