--- src/main/java/com/isn/quizplatform/exception/QuizException.java
@@ -1,5 +1,0 @@
-package com.isn.quizplatform.exception;
-
-public class QuizException {
-
-}
